pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coco-agent"
version = "2.0.0"
description = "CoCo v2 - AI Cohort Copilot for Healthcare RWD"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "Debu Sinha", email = "debusinha2009@gmail.com"}
]

dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "httpx>=0.27",
    "psycopg[binary]>=3.2",
    "psycopg_pool>=3.2",
    "jinja2>=3.1",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "databricks-sdk>=0.30",
    "databricks-vectorsearch>=0.40",
    "databricks-agents>=1.1",
    "mlflow[databricks]>=3.1",
    "dspy>=2.5",
    "pyarrow>=16",
    "pandas>=2.2",
    "sqlparse>=0.5",
    "python-jose[cryptography]>=3.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.14",
    "ruff>=0.4",
    "mypy>=1.8",
]
eval = [
    "openai>=1.10",
]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-select = ["E", "W", "F", "I"]
extend-ignore = ["E501"]

[tool.ruff.isort]
known-first-party = ["coco"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
no_implicit_optional = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: Unit tests (no external dependencies)",
    "integration: Integration tests (mocked Databricks services)",
    "slow: Slow tests (evaluation, data generation)",
]
addopts = "-v --tb=short"
